make io::stream module
