+++ src/io/mod.rs
@@ -3,4 +3,6 @@ pub mod functional;
 #[cfg(feature = "jpegturbo")]
 pub mod jpeg;
 #[cfg(feature = "gstreamer")]
+pub mod stream;
+#[cfg(feature = "gstreamer")]
 pub mod webcam;
